refactor: 코드 정리
- 사용하지 않는 임포트 문 정리
- 오타 정리
- 이제 필요 없는 fakeModule 삭제
- TODO 주석 구체적으로 작성
