+++ backend/src/auth/auth.controller.ts
@@ -1,11 +1,11 @@
 import { Body, Controller, Delete, Get, HttpCode, Post, Res, UseGuards } from '@nestjs/common';
-import { AuthService } from './auth.service';
-import { Response } from 'express';
 import { ConfigService } from '@nestjs/config';
-import { AccessTokenPayload, RefreshTokenPayload, TOKEN_LIFETIME_MAP } from './token/token.service';
-import { RefreshTokenGuard } from './guards/refreshToken.guard';
+import { Response } from 'express';
 import { User } from 'src/users/decorators/user.decorator';
+import { AuthService } from './auth.service';
 import { SkipAuthGuard } from './decorators/public.decorator';
+import { RefreshTokenGuard } from './guards/refreshToken.guard';
+import { AccessTokenPayload, RefreshTokenPayload, TOKEN_LIFETIME_MAP } from './token/token.service';
 
 export type OauthProvider = 'google' | 'kakao' | 'naver';
