+++ backend/src/auth/oauth/oauth.module.ts
@@ -1,5 +1,5 @@
-import { Module } from '@nestjs/common';
 import { HttpModule } from '@nestjs/axios';
+import { Module } from '@nestjs/common';
 import { GoogleService } from './google.service';
 import { KakaoService } from './kakao.service';
 import { NaverService } from './naver.service';
