+++ backend/src/daily-walk-time/daily-walk-time.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
-import { DailyWalkTimeService } from './daily-walk-time.service';
-import { DailyWalkTime } from './daily-walk-time.entity';
 import { TypeOrmModule } from '@nestjs/typeorm';
+import { DailyWalkTime } from './daily-walk-time.entity';
+import { DailyWalkTimeService } from './daily-walk-time.service';
 
 @Module({
     imports: [TypeOrmModule.forFeature([DailyWalkTime])],
