+++ backend/src/dog-walk-day/dog-walk-day.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
-import { DogWalkDayService } from './dog-walk-day.service';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { DogWalkDay } from './dog-walk-day.entity';
+import { DogWalkDayService } from './dog-walk-day.service';
 
 @Module({
     imports: [TypeOrmModule.forFeature([DogWalkDay])],
