+++ backend/src/dogs/dogs.controller.ts
@@ -1,10 +1,10 @@
 import { Controller, Get } from '@nestjs/common';
+import { AccessTokenPayload } from 'src/auth/token/token.service';
+import { Serialize } from 'src/common/interceptor/serialize.interceptor';
+import { User } from 'src/users/decorators/user.decorator';
 import { UsersService } from 'src/users/users.service';
 import { DogsService } from './dogs.service';
 import { DogStatisticDto } from './dto/dog-statistic.dto';
-import { Serialize } from 'src/common/interceptor/serialize.interceptor';
-import { User } from 'src/users/decorators/user.decorator';
-import { AccessTokenPayload } from 'src/auth/token/token.service';
 
 export type DogProfile = {
     id: number;
@@ -22,7 +22,7 @@ export class DogsController {
     @Get('/walk-available')
     async getAvailableDogs(@User() user: AccessTokenPayload): Promise<DogProfile[]> {
         const ownDogs = await this.usersService.getDogsList(user.userId);
-        return await this.dogsService.truncateNotAvaialableDog(ownDogs);
+        return await this.dogsService.truncateNotAvailableDog(ownDogs);
     }
 
     @Serialize(DogStatisticDto)
