+++ backend/src/entity/walk-log-photos.entity.ts
@@ -1,4 +1,4 @@
-import { Column, Entity, JoinColumn, ManyToOne, PrimaryColumn, PrimaryGeneratedColumn } from 'typeorm';
+import { Column, Entity, JoinColumn, ManyToOne, PrimaryColumn } from 'typeorm';
 import { WalkJournals } from '../journals/walk-journals.entity';
 
 @Entity('walk_log_photos')
