--- backend/src/fake/fake.module.ts
@@ -1,7 +1,0 @@
-import { Module } from '@nestjs/common';
-import { FakeController } from './fake.controller';
-
-@Module({
-    controllers: [FakeController],
-})
-export class FakeModule {}
