+++ backend/src/fixture/users.fixture.ts
@@ -1,5 +1,5 @@
-import { Users } from '../users/users.entity';
 import { Role } from '../users/user-roles.enum';
+import { Users } from '../users/users.entity';
 
 const ACCESS_TOKEN =
     'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOjEsImlhdCI6MTY2ODc1NzUzNCwiZXhwIjoxNjY4ODQzOTM0fQ.fwjmKJZ7enTKt7tPfNx-ZG_rczvhkz2ktMV5pDNbxkw';
