+++ backend/src/journals/journals.controller.ts
@@ -1,8 +1,12 @@
-import { Controller, Get } from '@nestjs/common';
+import { Controller, Post } from '@nestjs/common';
 import { JournalsService } from './journals.service';
-import { WalkJournals } from './walk-journals.entity';
 
 @Controller('journals')
 export class JournalsController {
     constructor(private readonly journalsService: JournalsService) {}
+
+    @Post('/')
+    fakeJournalsSave() {
+        return true;
+    }
 }
