+++ backend/src/journals/journals.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common';
-import { JournalsService } from './journals.service';
-import { JournalsController } from './journals.controller';
-import { WalkJournalsRepository } from './walk-jorunals.repository';
 import { DatabaseModule } from 'src/common/database/database.module';
+import { JournalsController } from './journals.controller';
+import { JournalsService } from './journals.service';
 import { WalkJournals } from './walk-journals.entity';
+import { WalkJournalsRepository } from './walk-journals.repository';
 
 @Module({
     imports: [DatabaseModule.forFeature([WalkJournals])],
