+++ backend/src/journals/walk-journals.entity.ts
@@ -1,5 +1,5 @@
-import { Users } from '../users/users.entity';
 import { Column, Entity, JoinColumn, ManyToOne, PrimaryGeneratedColumn } from 'typeorm';
+import { Users } from '../users/users.entity';
 
 @Entity('walk_journals')
 export class WalkJournals {
