+++ backend/src/users/user-dogs.entity.ts
@@ -1,6 +1,6 @@
 import { Entity, JoinColumn, ManyToOne, PrimaryColumn } from 'typeorm';
-import { Users } from './users.entity';
 import { Dogs } from '../dogs/dogs.entity';
+import { Users } from './users.entity';
 
 @Entity('users_dogs')
 export class UsersDogs {
