+++ backend/src/users/users.entity.ts
@@ -1,6 +1,6 @@
+import { Dogs } from 'src/dogs/dogs.entity';
 import { Column, CreateDateColumn, Entity, JoinColumn, ManyToOne, PrimaryGeneratedColumn } from 'typeorm';
 import { Role } from './user-roles.enum';
-import { Dogs } from 'src/dogs/dogs.entity';
 
 @Entity()
 export class Users {
