+++ backend/src/walk/dtos/walk-command.dto.ts
@@ -1,4 +1,4 @@
-import { IsArray, IsInt } from 'class-validator';
+import { IsArray } from 'class-validator';
 
 export class WalkCommandDto {
     @IsArray()
